Refactor AddEdit countdown view into modular subviews
