--- CouplesCount/Views/AddEditCountdownView.swift
@@ -1,568 +1,0 @@
-import SwiftUI
-import SwiftData
-import PhotosUI
-import AVFoundation
-import UIKit
-
-// MARK: - Reminder options
-
-enum ReminderOption: Int, CaseIterable, Identifiable {
-    case h1 = -60
-    case h2 = -120
-    case h3 = -180
-    case h6 = -360
-    case h12 = -720
-    case d1 = -1440
-    case d2 = -2880
-    case d3 = -4320
-    case d7 = -10080
-
-    var id: Int { rawValue }
-    var label: String {
-        switch self {
-        case .h1: return "1h"
-        case .h2: return "2h"
-        case .h3: return "3h"
-        case .h6: return "6h"
-        case .h12: return "12h"
-        case .d1: return "1d"
-        case .d2: return "2d"
-        case .d3: return "3d"
-        case .d7: return "7d"
-        }
-    }
-}
-
-// MARK: - Add/Edit Screen
-
-struct AddEditCountdownView: View {
-    @Environment(\.dismiss) private var dismiss
-    @Environment(\.modelContext) private var modelContext
-    @EnvironmentObject private var theme: ThemeManager
-
-    let existing: Countdown?
-
-    @State private var title: String = ""
-    @State private var date: Date = Date().addingTimeInterval(86_400)
-    @State private var timeZoneID: String = TimeZone.current.identifier
-    @State private var cardFontStyle: CardFontStyle = .classic
-
-    // Background selection
-    @State private var backgroundStyle: String = "color" // "color" | "image"
-    @State private var colorHex: String = ""
-    @State private var imageData: Data? = nil
-    @State private var showPhotoPicker = false
-    @State private var showCamera = false
-
-    // Live preview values
-    @State private var previewTitle: String = "Countdown"
-    @State private var previewDate: Date = Date().addingTimeInterval(86_400)
-    @State private var previewColorHex: String = ""
-    @State private var previewImageData: Data? = nil
-
-    // Reminders
-    @State private var selectedReminders: Set<ReminderOption> = []
-    @State private var showReminderSheet = false
-
-    // UX
-    @State private var showValidation = false
-    @State private var saveError: String?
-
-    // Sharing
-    @State private var isShared: Bool = false
-    @State private var selectedFriends: Set<UUID> = []
-    @Query(sort: \Friend.name) private var friends: [Friend]
-    @State private var shareURL: URL? = nil
-    @State private var showShareSheet = false
-
-    init(existing: Countdown? = nil) { self.existing = existing }
-
-    var body: some View {
-        NavigationStack {
-            ScrollView {
-                VStack(spacing: 18) {
-
-                    // MARK: Swipable widget preview (square → rectangular)
-                    TabView {
-                        // Square
-                        WidgetPreview(
-                            title: previewTitle,
-                            targetDate: previewDate,
-                            tzID: timeZoneID,
-                            style: cardFontStyle,
-                            backgroundStyle: backgroundStyle,
-                            bgColorHex: previewColorHex,
-                            imageData: previewImageData
-                        )
-                        .frame(width: 160, height: 160)
-                        .clipShape(RoundedRectangle(cornerRadius: 22, style: .continuous))
-                        .padding(.vertical, 8)
-
-                        // Rectangular
-                        WidgetPreview(
-                            title: previewTitle,
-                            targetDate: previewDate,
-                            tzID: timeZoneID,
-                            style: cardFontStyle,
-                            backgroundStyle: backgroundStyle,
-                            bgColorHex: previewColorHex,
-                            imageData: previewImageData
-                        )
-                        .frame(height: 140)
-                        .clipShape(RoundedRectangle(cornerRadius: 22, style: .continuous))
-                        .padding(.vertical, 8)
-                    }
-                    .tabViewStyle(.page)
-                    .indexViewStyle(.page(backgroundDisplayMode: .automatic))
-                    .frame(height: 180)
-                    .padding(.horizontal, 16)
-                    .animation(.easeInOut(duration: 0.2), value: title)
-                    .animation(.easeInOut(duration: 0.2), value: date)
-                    .animation(.easeInOut(duration: 0.2), value: backgroundStyle)
-                    .animation(.easeInOut(duration: 0.2), value: colorHex)
-                    .animation(.easeInOut(duration: 0.2), value: imageData)
-
-                    // MARK: Details
-                    SettingsCard {
-                        TextField("", text: $title, prompt: Text("Title (e.g., Anniversary)" ).foregroundStyle(theme.theme.textSecondary))
-                            .foregroundStyle(theme.theme.textPrimary)
-                            .textInputAutocapitalization(.words)
-                            .onSubmit { lightHaptic() }
-
-                        Picker("Font", selection: $cardFontStyle) {
-                            ForEach(CardFontStyle.allCases) { f in
-                                Text(f.displayName).tag(f)
-                            }
-                        }
-                        .pickerStyle(.segmented)
-                        .onChange(of: cardFontStyle, initial: false) { _, _ in lightHaptic() }
-
-                        HStack {
-                            DatePicker("Date", selection: $date, displayedComponents: .date)
-                                .foregroundStyle(theme.theme.textPrimary)
-                            DatePicker("", selection: $date, displayedComponents: .hourAndMinute)
-                                .labelsHidden()
-                                .foregroundStyle(theme.theme.textPrimary)
-                        }
-
-                        NavigationLink {
-                            TimeZonePickerView(selectedID: $timeZoneID)
-                        } label: {
-                            HStack {
-                                Text("Time Zone")
-                                    .foregroundStyle(theme.theme.textPrimary)
-                                Spacer()
-                                Text(TimeZone(identifier: timeZoneID)?.identifier ?? "System")
-                                    .foregroundStyle(theme.theme.textSecondary)
-                            }
-                        }
-                    }
-
-                    // MARK: Background
-                    SettingsCard {
-                        Text("Background")
-                            .font(.subheadline.weight(.semibold))
-                            .foregroundStyle(theme.theme.textSecondary)
-
-                        Picker("Style", selection: $backgroundStyle) {
-                            Text("Color").tag("color")
-                            Text("Image").tag("image")
-                        }
-                        .pickerStyle(.segmented)
-
-                        if backgroundStyle == "color" {
-                            HStack(spacing: 10) {
-                                ForEach(["#0A84FF","#5856D6","#FF2D55","#34C759","#FF9F0A"], id: \.self) { hex in
-                                    Circle()
-                                        .fill(Color(hex: hex) ?? .blue)
-                                        .frame(width: 32, height: 32)
-                                        .overlay(
-                                            Circle().stroke(theme.theme.textPrimary.opacity(colorHex == hex ? 0.9 : 0), lineWidth: 2)
-                                        )
-                                        .onTapGesture { colorHex = hex }
-                                }
-                                Spacer()
-                                ColorPicker("", selection: Binding(
-                                    get: { Color(hex: colorHex) ?? .blue },
-                                    set: { colorHex = $0.hexString }
-                                ))
-                                .labelsHidden()
-                            }
-                        } else {
-                            if let data = imageData, let ui = UIImage(data: data) {
-                                Image(uiImage: ui)
-                                    .resizable()
-                                    .scaledToFill()
-                                    .frame(height: 140)
-                                    .clipShape(RoundedRectangle(cornerRadius: 12, style: .continuous))
-                                    .accessibilityHidden(true)
-                            } else {
-                                Text("No image selected")
-                                    .foregroundStyle(theme.theme.textSecondary)
-                            }
-
-                            HStack(spacing: 12) {
-                                Button {
-                                    PHPhotoLibrary.requestAuthorization { status in
-                                        if status == .authorized || status == .limited {
-                                            DispatchQueue.main.async { showPhotoPicker = true }
-                                        }
-                                    }
-                                } label: { labelButton("Choose Photo", system: "photo") }
-
-                                Button {
-                                    AVCaptureDevice.requestAccess(for: .video) { ok in
-                                        if ok { DispatchQueue.main.async { showCamera = true } }
-                                    }
-                                } label: { labelButton("Camera", system: "camera") }
-                            }
-                        }
-                    }
-                    // MARK: Sharing
-                    SettingsCard {
-                        Toggle("Shared countdown", isOn: $isShared)
-                            .foregroundStyle(theme.theme.textPrimary)
-                        if isShared {
-                            ForEach(friends) { friend in
-                                let isSelected = Binding<Bool>(
-                                    get: { selectedFriends.contains(friend.id) },
-                                    set: { newVal in
-                                        if newVal { selectedFriends.insert(friend.id) } else { selectedFriends.remove(friend.id) }
-                                    }
-                                )
-                                Toggle(friend.name, isOn: isSelected)
-                                    .foregroundStyle(theme.theme.textPrimary)
-                            }
-                        }
-                    }
-
-                    // MARK: Reminders
-                    SettingsCard {
-                        HStack {
-                            Text("Reminders")
-                                .font(.subheadline.weight(.semibold))
-                                .foregroundStyle(theme.theme.textSecondary)
-                            Spacer()
-                            Button("+ Add Reminder") {
-                                NotificationManager.requestAuthorizationIfNeeded()
-                                showReminderSheet = true
-                            }
-                            .foregroundStyle(theme.theme.textPrimary)
-                        }
-
-                        if !selectedReminders.isEmpty {
-                            LazyVGrid(columns: [GridItem(.adaptive(minimum: 60), spacing: 8)], alignment: .leading, spacing: 8) {
-                                ForEach(Array(selectedReminders).sorted { $0.rawValue < $1.rawValue }, id: \.self) { opt in
-                                    HStack(spacing: 4) {
-                                        Text(opt.label)
-                                            .foregroundStyle(theme.theme.textPrimary)
-                                        Image(systemName: "xmark.circle.fill")
-                                            .foregroundStyle(theme.theme.textSecondary)
-                                            .onTapGesture { selectedReminders.remove(opt) }
-                                    }
-                                    .padding(.horizontal, 8)
-                                    .padding(.vertical, 4)
-                                    .background(theme.theme.textPrimary.opacity(0.1))
-                                    .clipShape(Capsule())
-                                }
-                            }
-                            .padding(.top, 4)
-                        }
-                    }
-                    .sheet(isPresented: $showReminderSheet) {
-                        ReminderPicker(selections: $selectedReminders)
-                            .environmentObject(theme)
-                    }
-
-                    if showValidation && title.trimmingCharacters(in: .whitespaces).isEmpty {
-                        Text("Please enter a title.")
-                            .foregroundStyle(.red)
-                            .padding(.horizontal, 16)
-                    }
-
-                    if let existing {
-                        SettingsCard {
-                            Button {
-                                withAnimation(.spring(response: 0.4, dampingFraction: 0.85)) {
-                                    existing.isArchived.toggle()
-                                    if existing.isArchived {
-                                        NotificationManager.cancelReminders(for: existing.id)
-                                    }
-                                    try? modelContext.save()
-                                    let all = (try? modelContext.fetch(FetchDescriptor<Countdown>())) ?? []
-                                    updateWidgetSnapshot(afterSaving: all)
-                                }
-                                if existing.isArchived { Haptics.light() }
-                                dismiss()
-                            } label: {
-                                Label(existing.isArchived ? "Unarchive Countdown" : "Archive Countdown",
-                                      systemImage: existing.isArchived ? "tray.and.arrow.up" : "archivebox")
-                                    .foregroundStyle(theme.theme.textPrimary)
-                            }
-                        }
-
-                        SettingsCard {
-                            Button(role: .destructive) {
-                                withAnimation(.spring(response: 0.4, dampingFraction: 0.85)) {
-                                    NotificationManager.cancelAll(for: existing.id)
-                                    modelContext.delete(existing)
-                                    try? modelContext.save()
-                                    let all = (try? modelContext.fetch(FetchDescriptor<Countdown>())) ?? []
-                                    updateWidgetSnapshot(afterSaving: all)
-                                }
-                                Haptics.warning()
-                                dismiss()
-                            } label: {
-                                Label("Delete Countdown", systemImage: "trash")
-                                    .foregroundStyle(theme.theme.textPrimary)
-                            }
-                        }
-                    }
-                }
-                .padding(.top, 14)
-                .padding(.horizontal, 16)
-            }
-            .scrollIndicators(.hidden)
-            .overlay(alignment: .trailing) {
-                RoundedRectangle(cornerRadius: 3)
-                    .fill(.gray.opacity(0.4))
-                    .frame(width: 6)
-                    .padding(.vertical, 8)
-                    .padding(.trailing, 2)
-            }
-            .background(theme.theme.background.ignoresSafeArea())
-            .tint(theme.theme.textPrimary)
-            .navigationTitle(existing == nil ? "Add Countdown" : "Edit Countdown")
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbarColorScheme(theme.theme == .light ? .light : .dark, for: .navigationBar)
-            .toolbarBackground(theme.theme.background, for: .navigationBar)
-            .toolbarBackground(.visible, for: .navigationBar)
-            .toolbar {
-                ToolbarItem(placement: .cancellationAction) {
-                    Button("Cancel") { dismiss() }
-                        .foregroundStyle(theme.theme.textPrimary)
-                }
-                ToolbarItem(placement: .principal) {
-                    Text(existing == nil ? "Add Countdown" : "Edit Countdown")
-                        .foregroundStyle(theme.theme.textPrimary)
-                }
-                ToolbarItem(placement: .navigationBarTrailing) {
-                    HStack {
-                        if let existing {
-                            Button {
-                                shareURL = CountdownShareService.exportURL(for: existing)
-                                showShareSheet = shareURL != nil
-                            } label: {
-                                Image(systemName: "square.and.arrow.up")
-                                    .frame(width: 44, height: 44)
-                                    .contentShape(Rectangle())
-                                    .accessibilityLabel("Share")
-                                    .accessibilityHint("Share countdown")
-                                    .foregroundStyle(theme.theme.textPrimary)
-                            }
-                        }
-                        Button(action: save) {
-                            Image(systemName: "checkmark")
-                                .frame(width: 44, height: 44)
-                                .contentShape(Rectangle())
-                                .accessibilityLabel("Save")
-                                .accessibilityHint("Save countdown")
-                                .foregroundStyle(theme.theme.textPrimary)
-                        }
-                        .disabled(title.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty)
-                    }
-                }
-            }
-            .sheet(isPresented: $showPhotoPicker) { PhotoPicker(imageData: $imageData) }
-            .sheet(isPresented: $showCamera) { CameraPicker(imageData: $imageData) }
-            .sheet(isPresented: $showShareSheet) {
-                if let shareURL {
-                    ShareSheet(activityItems: [shareURL])
-                }
-            }
-            // Live preview & haptics
-            .onChange(of: title, initial: false) { _, new in
-                previewTitle = new.isEmpty ? "Countdown" : new
-            }
-            .onChange(of: date, initial: false) { _, new in
-                previewDate = new
-            }
-            .onChange(of: colorHex, initial: false) { _, new in
-                previewColorHex = new
-                lightHaptic()
-            }
-            .onChange(of: imageData, initial: false) { _, new in
-                previewImageData = new
-                if new != nil { lightHaptic() }
-            }
-            .onAppear {
-                let defaultHex = theme.theme.primary.hexString
-                if let existing {
-                    title = existing.title
-                    date = existing.targetDate
-                    timeZoneID = existing.timeZoneID
-                    cardFontStyle = existing.cardFontStyle
-                    backgroundStyle = existing.backgroundStyle
-                    let stored = existing.backgroundColorHex?.uppercased()
-                    if stored == nil || stored == "#FFFFFF" || stored == defaultHex.uppercased() {
-                        colorHex = defaultHex
-                    } else {
-                        colorHex = stored ?? defaultHex
-                    }
-                    imageData = existing.backgroundImageData
-                    selectedReminders = Set(existing.reminderOffsets.compactMap { ReminderOption(rawValue: $0) })
-                    isShared = existing.isShared
-                    selectedFriends = Set(existing.sharedWith.map { $0.id })
-                    previewTitle = existing.title
-                    previewDate = existing.targetDate
-                    previewColorHex = colorHex
-                    previewImageData = existing.backgroundImageData
-                } else {
-                    NotificationManager.requestAuthorizationIfNeeded()
-                    colorHex = defaultHex
-                    previewTitle = title
-                    previewDate = date
-                    previewColorHex = defaultHex
-                    previewImageData = imageData
-                }
-            }
-        }
-        .tint(theme.theme.textPrimary)
-        .alert("Couldn’t Save",
-               isPresented: Binding(get: { saveError != nil },
-                                   set: { if !$0 { saveError = nil } })) {
-            Button("OK", role: .cancel) { }
-        } message: {
-            Text(saveError ?? "Unknown error")
-        }
-    }
-
-    // MARK: Save
-
-    private func save() {
-        let trimmed = title.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty else { showValidation = true; return }
-
-        do {
-            let defaultHex = theme.theme.primary.hexString.uppercased()
-            let chosenHex = colorHex.uppercased()
-            let storedHex: String? = (chosenHex == defaultHex) ? nil : chosenHex
-
-            if let existing {
-                existing.title = trimmed
-                existing.targetDate = date
-                existing.timeZoneID = timeZoneID
-                existing.cardFontStyle = cardFontStyle
-                existing.backgroundStyle = backgroundStyle
-                existing.backgroundColorHex = storedHex
-                existing.backgroundImageData = imageData
-                existing.hasImage = backgroundStyle == "image"
-                existing.reminderOffsets = selectedReminders.map { $0.rawValue }
-                existing.lastEdited = .now
-                existing.isShared = isShared
-                existing.sharedWith = friends.filter { selectedFriends.contains($0.id) }
-                NotificationManager.cancelAll(for: existing.id)
-                if !selectedReminders.isEmpty { NotificationManager.scheduleReminders(for: existing) }
-            } else {
-                let cd = Countdown(
-                    title: trimmed,
-                    targetDate: date,
-                    timeZoneID: timeZoneID,
-                    cardFontStyle: cardFontStyle,
-                    backgroundStyle: backgroundStyle,
-                    backgroundColorHex: storedHex,
-                    backgroundImageData: imageData,
-                    reminderOffsets: selectedReminders.map { $0.rawValue },
-                    isShared: isShared,
-                    sharedWith: friends.filter { selectedFriends.contains($0.id) }
-                )
-                withAnimation(.spring(response: 0.4, dampingFraction: 0.85)) {
-                    modelContext.insert(cd)
-                }
-                if !selectedReminders.isEmpty { NotificationManager.scheduleReminders(for: cd) }
-            }
-
-            try modelContext.save()
-            let all = try modelContext.fetch(FetchDescriptor<Countdown>())
-            updateWidgetSnapshot(afterSaving: all)
-            dismiss()
-        } catch {
-            saveError = error.localizedDescription
-        }
-    }
-
-    @ViewBuilder
-    private func labelButton(_ title: String, system: String) -> some View {
-        Label(title, systemImage: system)
-            .foregroundStyle(theme.theme.textPrimary)
-            .padding(.vertical, 8)
-            .padding(.horizontal, 12)
-            .background(
-                RoundedRectangle(cornerRadius: 10, style: .continuous)
-                    .fill(theme.theme.textPrimary.opacity(0.1))
-            )
-    }
-
-    private func lightHaptic() {
-        UIImpactFeedbackGenerator(style: .light).impactOccurred()
-    }
-}
-
-// MARK: - Reminder Picker Sheet
-
-struct ReminderPicker: View {
-    @Environment(\.dismiss) private var dismiss
-    @EnvironmentObject private var theme: ThemeManager
-    @Binding var selections: Set<ReminderOption>
-    @State private var temp: Set<ReminderOption>
-
-    init(selections: Binding<Set<ReminderOption>>) {
-        self._selections = selections
-        self._temp = State(initialValue: selections.wrappedValue)
-    }
-
-    var body: some View {
-        NavigationStack {
-            ScrollView {
-                LazyVGrid(columns: [GridItem(.adaptive(minimum: 60), spacing: 12)], spacing: 12) {
-                    ForEach(ReminderOption.allCases) { option in
-                        let isSel = temp.contains(option)
-                        Text(option.label)
-                            .foregroundStyle(theme.theme.textPrimary)
-                            .frame(maxWidth: .infinity, minHeight: 44)
-                            .background(isSel ? theme.theme.textPrimary.opacity(0.2) : theme.theme.textPrimary.opacity(0.1))
-                            .clipShape(RoundedRectangle(cornerRadius: 8, style: .continuous))
-                            .overlay(
-                                RoundedRectangle(cornerRadius: 8)
-                                    .stroke(isSel ? theme.theme.textPrimary : .clear, lineWidth: 2)
-                            )
-                            .onTapGesture {
-                                if isSel { temp.remove(option) } else { temp.insert(option) }
-                            }
-                    }
-                }
-                .padding()
-            }
-            .background(theme.theme.background.ignoresSafeArea())
-            .tint(theme.theme.textPrimary)
-            .navigationTitle("Reminders")
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbarColorScheme(theme.theme == .light ? .light : .dark, for: .navigationBar)
-            .toolbarBackground(theme.theme.background, for: .navigationBar)
-            .toolbarBackground(.visible, for: .navigationBar)
-            .toolbar {
-                ToolbarItem(placement: .principal) {
-                    Text("Reminders")
-                        .foregroundStyle(theme.theme.textPrimary)
-                }
-                ToolbarItem(placement: .confirmationAction) {
-                    Button("Add") {
-                        selections = temp
-                        dismiss()
-                    }
-                    .foregroundStyle(theme.theme.textPrimary)
-                }
-            }
-        }
-    }
-}
