+++ CouplesCount/Views/Countdowns/AddEdit/BackgroundPickerSection.swift
@@ -1,0 +1,91 @@
+import SwiftUI
+import PhotosUI
+import AVFoundation
+import UIKit
+
+struct BackgroundPickerSection: View {
+    @EnvironmentObject private var theme: ThemeManager
+    @Binding var backgroundStyle: String
+    @Binding var colorHex: String
+    @Binding var imageData: Data?
+
+    @State private var showPhotoPicker = false
+    @State private var showCamera = false
+
+    var body: some View {
+        SettingsCard {
+            Text("Background")
+                .font(.subheadline.weight(.semibold))
+                .foregroundStyle(theme.theme.textSecondary)
+
+            Picker("Style", selection: $backgroundStyle) {
+                Text("Color").tag("color")
+                Text("Image").tag("image")
+            }
+            .pickerStyle(.segmented)
+
+            if backgroundStyle == "color" {
+                HStack(spacing: 10) {
+                    ForEach(["#0A84FF","#5856D6","#FF2D55","#34C759","#FF9F0A"], id: \.self) { hex in
+                        Circle()
+                            .fill(Color(hex: hex) ?? .blue)
+                            .frame(width: 32, height: 32)
+                            .overlay(
+                                Circle().stroke(theme.theme.textPrimary.opacity(colorHex == hex ? 0.9 : 0), lineWidth: 2)
+                            )
+                            .onTapGesture { colorHex = hex }
+                    }
+                    Spacer()
+                    ColorPicker("", selection: Binding(
+                        get: { Color(hex: colorHex) ?? .blue },
+                        set: { colorHex = $0.hexString }
+                    ))
+                    .labelsHidden()
+                }
+            } else {
+                if let data = imageData, let ui = UIImage(data: data) {
+                    Image(uiImage: ui)
+                        .resizable()
+                        .scaledToFill()
+                        .frame(height: 140)
+                        .clipShape(RoundedRectangle(cornerRadius: 12, style: .continuous))
+                        .accessibilityHidden(true)
+                } else {
+                    Text("No image selected")
+                        .foregroundStyle(theme.theme.textSecondary)
+                }
+
+                HStack(spacing: 12) {
+                    Button {
+                        PHPhotoLibrary.requestAuthorization { status in
+                            if status == .authorized || status == .limited {
+                                DispatchQueue.main.async { showPhotoPicker = true }
+                            }
+                        }
+                    } label: { labelButton("Choose Photo", system: "photo") }
+
+                    Button {
+                        AVCaptureDevice.requestAccess(for: .video) { ok in
+                            if ok { DispatchQueue.main.async { showCamera = true } }
+                        }
+                    } label: { labelButton("Camera", system: "camera") }
+                }
+            }
+        }
+        .sheet(isPresented: $showPhotoPicker) { PhotoPicker(imageData: $imageData) }
+        .sheet(isPresented: $showCamera) { CameraPicker(imageData: $imageData) }
+    }
+
+    @ViewBuilder
+    private func labelButton(_ title: String, system: String) -> some View {
+        Label(title, systemImage: system)
+            .foregroundStyle(theme.theme.textPrimary)
+            .padding(.vertical, 8)
+            .padding(.horizontal, 12)
+            .background(
+                RoundedRectangle(cornerRadius: 10, style: .continuous)
+                    .fill(theme.theme.textPrimary.opacity(0.1))
+            )
+    }
+}
+
